Stop throwing exceptions.
The UnknownPropertyException being thrown whenever a non-existent
property or method was requested caused a huge performance hit.

Instead the library now returns null by default. This only affects
method calls against the presenter which don't exist on the presenter or
underlying model - null will be returned instead of throwing an
exception.

The only downside to this is a method you intended to define on a
presenter but never did will not be brought to your attention with an
exception when attempting to access the missing method.